tickからMBSに変換する処理をdomain.tsに移動
