+++ src/components/Sing/PlayheadPositionDisplay.vue
@@ -20,12 +20,8 @@ import { useStore } from "@/store";
 import ContextMenu, {
   ContextMenuItemData,
 } from "@/components/Menu/ContextMenu.vue";
-import {
-  getBeatDuration,
-  getMeasureDuration,
-  getTimeSignaturePositions,
-} from "@/sing/domain";
-import { TimeSignature } from "@/store/type";
+import { getTimeSignaturePositions, tickToMbs } from "@/sing/domain";
+import { MBS } from "@/store/type";
 
 const store = useStore();
 
@@ -42,51 +38,12 @@ const timeSignatures = computed(() => {
   }));
 });
 
-const findTimeSignatureIndex = (
-  ticks: number,
-  timeSignatures: (TimeSignature & { position: number })[],
-) => {
-  if (ticks < 0) {
-    return 0;
-  }
-  for (let i = 0; i < timeSignatures.length - 1; i++) {
-    if (
-      timeSignatures[i].position <= ticks &&
-      timeSignatures[i + 1].position > ticks
-    ) {
-      return i;
-    }
-  }
-  return timeSignatures.length - 1;
-};
-
-// Measures, Beats, Sixteenths
-const mbs = computed(() => {
+const mbs = computed((): MBS => {
   if (displayMode.value !== "MBS") {
-    return { measures: 1, beats: 1, sixteenths: 1, remaining: 0 };
+    return { measures: 1, beats: 1, sixteenths: 1 };
   }
   const tpqn = store.state.tpqn;
-
-  const ticks = playheadTicks.value;
-
-  const tsIndex = findTimeSignatureIndex(ticks, timeSignatures.value);
-  const ts = timeSignatures.value[tsIndex];
-
-  const measureDuration = getMeasureDuration(ts.beats, ts.beatType, tpqn);
-  const beatDuration = getBeatDuration(ts.beats, ts.beatType, tpqn);
-  const sixteenthDuration = tpqn / 4;
-
-  const posInTs = ticks - ts.position;
-  const measuresInTs = Math.floor(posInTs / measureDuration);
-  const measures = ts.measureNumber + measuresInTs;
-
-  const posInMeasure = posInTs - measureDuration * measuresInTs;
-  const beats = 1 + Math.floor(posInMeasure / beatDuration);
-
-  const posInBeat = posInMeasure - beatDuration * (beats - 1);
-  const sixteenths = 1 + posInBeat / sixteenthDuration;
-
-  return { measures, beats, sixteenths };
+  return tickToMbs(playheadTicks.value, timeSignatures.value, tpqn);
 });
 
 const measuresStr = computed(() => {
